pyproject.toml
[project]
name = "rag-agent"
version = "0.1.0"
description = ""
authors = [
    {name = "Hamza Hassan",email = "hamza_hassan45@hotmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.12"

[tool.poetry.dependencies]
accelerate = "0.28.0"
aiohttp = "3.9.3"
aiosignal = "1.3.1"
annotated-types = "0.6.0"
anyio = "4.3.0"
attrs = "23.2.0"
beautifulsoup4 = "4.12.3"
bs4 = "0.0.2"
certifi = "2024.2.2"
charset-normalizer = "3.3.2"
click = "8.1.7"
dataclasses-json = "0.6.4"
Deprecated = "1.2.14"
dirtyjson = "1.0.8"
diskcache = "5.6.3"
distro = "1.9.0"
fastapi = "0.110.1"
filelock = "3.13.3"
frozenlist = "1.4.1"
fsspec = "2024.3.1"
greenlet = "3.0.3"
guidance = "0.1.13"
h11 = "0.14.0"
html2text = "2024.2.26"
httpcore = "1.0.5"
httpx = "0.27.0"
huggingface-hub = "0.20.3"
idna = "3.6"
interegular = "0.3.3"
Jinja2 = "3.1.3"
joblib = "1.3.2"
llama-hub = "0.0.79.post1"
llama-index = "0.10.25"
llama-index-agent-openai = "0.2.1"
llama-index-cli = "0.1.11"
llama-index-core = "0.10.25.post2"
llama-index-embeddings-huggingface = "0.2.0"
llama-index-embeddings-openai = "0.1.7"
llama-index-indices-managed-llama-cloud = "0.1.5"
llama-index-legacy = "0.9.48"
llama-index-llms-huggingface = "0.1.4"
llama-index-llms-llama-cpp = "0.1.3"
llama-index-llms-ollama = "0.1.2"
llama-index-llms-openai = "0.1.13"
llama-index-multi-modal-llms-openai = "0.1.4"
llama-index-program-guidance = "0.1.2"
llama-index-program-lmformatenforcer = "0.1.2"
llama-index-program-openai = "0.1.5"
llama-index-question-gen-openai = "0.1.3"
llama-index-readers-file = "0.1.12"
llama-index-readers-llama-parse = "0.1.4"
llama-parse = "0.4.0"
llama_cpp_python = "0.2.58"
llamaindex-py-client = "0.1.15"
lm-format-enforcer = "0.9.3"
MarkupSafe = "2.1.5"
marshmallow = "3.21.1"
minijinja = "1.0.16"
mpmath = "1.3.0"
multidict = "6.0.5"
mypy-extensions = "1.0.0"
nest-asyncio = "1.6.0"
networkx = "3.2.1"
nltk = "3.8.1"
numpy = "1.26.4"
openai = "1.14.3"
ordered-set = "4.1.0"
packaging = "24.0"
pandas = "2.2.1"
pillow = "10.2.0"
platformdirs = "4.2.0"
protobuf = "5.26.1"
psutil = "5.9.8"
pyaml = "23.12.0"
pydantic = "2.6.4"
pydantic_core = "2.16.3"
pydot = "2.0.0"
pyformlang = "1.0.9"
PyMuPDF = "1.24.0"
PyMuPDFb = "1.24.0"
pyparsing = "3.1.2"
pypdf = "4.1.0"
python-dateutil = "2.9.0.post0"
python-dotenv = "1.0.1"
pytz = "2024.1"
PyYAML = "6.0.1"
regex = "2023.12.25"
requests = "2.31.0"
retrying = "1.3.4"
safetensors = "0.4.2"
scikit-learn = "1.4.1.post1"
scipy = "1.12.0"
sentence-transformers = "2.6.1"
six = "1.16.0"
sniffio = "1.3.1"
soupsieve = "2.5"
SQLAlchemy = "2.0.29"
starlette = "0.37.2"
striprtf = "0.0.26"
sympy = "1.12"
tenacity = "8.2.3"
threadpoolctl = "3.4.0"
tiktoken = "0.6.0"
tokenizers = "0.15.2"
torch = "2.2.2"
tqdm = "4.66.2"
transformers = "4.39.2"
typing-inspect = "0.9.0"
typing_extensions = "4.10.0"
tzdata = "2024.1"
urllib3 = "2.2.1"
uvicorn = "0.29.0"
wrapt = "1.16.0"
yarl = "1.9.4"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
